requirements.txt
beautifulsoup4==4.12.2
matplotlib==3.8.0rc1
numpy==1.23.5
pandas==2.0.3
Pillow>=9.3.0
Pyrebase4==4.7.1
Requests<=2.31.0
scikit_learn==1.3.2
streamlit==1.28.1
streamlit_lottie==0.0.5
urllib3==1.26.15